[TASK] Introduced UrlService and LocaleService to prevent duplicate code (#304)
* [TASK] Introduced UrlService and LocaleService to prevent duplicate code

* [TASK] Remove unused imports from SnippetPreview

* [TASK] Removed unused constants from ModuleController

* [TASK] Revert removal of FE_PREVIEW_TYPE within PageLayoutHeader and added deprecated comment
